commit first time tho already created js file.
Finished creating the game in Javascript and tested in console.
Every steps are explained with comments in js. I tried to write the comments clear in plain English.
Created some input and output elements for game in html but not yet linked with js.
Have a css file too to style the page later....