Link the game with the DOM
The game is now linked with page and works well.
Change the method of getting input from player from window.prompt to buttons in page.
The player can now conveniently play the game in normal user interface.
The for loop worked well with the prompt method of getting input but it doesn't work as expected with dom input buttons, so I removed it.

The game is now playable but still need some improvements like when we reach total game result limit of 5, showing play again option.
For now, user can still play even if he or computer has won 5 matches.